初步完成MessageFragment

+++ app/src/main/java/top/littlefogcat/clickerx/model/entities/Config.kt
@@ -5,12 +5,14 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 /**
+ *
+ *
  * @Author littlefogcat
  * @Date 2020/8/4-1:13
  * @Email littlefogcat@foxmail.com
  */
 @Entity
 data class Config(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     @ColumnInfo(name = "title") val title: String
 )
